# public/swagger.yaml
openapi: 3.0.3
info:
  title: Air5Star eCommerce AI-Powered API
  description: |
    AI-powered features API for Air5Star HVAC eCommerce platform.
    
    This API provides intelligent features including:
    - Smart chatbot integration with product lookup
    - Advanced search with autocomplete and highlighting
    - Personalized product recommendations
    - Abandoned cart email automation
    - Inventory forecasting and alerts
    
    ## Security
    - All endpoints use HTTPS
    - API keys stored in environment variables
    - HMAC signature verification for webhooks
    - Minimal PII exposure to third-party services
    
    ## Rate Limiting
    - Search endpoints: 100 requests/minute
    - Recommendation endpoints: 50 requests/minute
    - Webhook endpoints: 1000 requests/hour
  version: 1.0.0
  contact:
    name: Air5Star Development Team
    email: dev@air5star.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://air5star.com/api
    description: Production server
paths:
...
  /search:
    get:
      tags:
        - Search
      summary: Smart search with autocomplete
      description: |
        Intelligent product search with autocomplete, highlighting, and relevance scoring.
        Supports both full search and autocomplete modes.
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
            maxLength: 100
          example: "split ac"
        - name: autocomplete
          in: query
          description: Enable autocomplete mode
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: category
          in: query
          description: Filter by product category
          schema:
            type: string
          example: "Air Conditioners"
        - name: min_price
          in: query
          description: Minimum price filter
          schema:
            type: number
            minimum: 0
        - name: max_price
          in: query
          description: Maximum price filter
          schema:
            type: number
            minimum: 0
        - name: offset
          in: query
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort_by
          in: query
          description: Sort results by specified field
          schema:
            type: string
            enum: ["relevance", "price_asc", "price_desc", "name"]
            default: "relevance"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              example:
                success: true
                data:
                  results:
                    - id: "1"
                      name: "Split AC 1.5 Ton Inverter"
                      category: "Air Conditioners"
                      price: 35000
                      imageUrl: "/images/split-ac-1.jpg"
                      relevance_score: 0.95
                      highlighted_name: "<mark>Split AC</mark> 1.5 Ton Inverter"
                  suggestions:
                    - "split ac 1 ton"
                    - "split ac 2 ton"
                  total: 15
                  query: "split ac"
                  autocomplete: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    HmacSignature:
      type: apiKey
      in: header
      name: X-Tidio-Signature
      description: HMAC SHA-256 signature for webhook verification
  schemas:
...
    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            results:
              type: array
              items:
                $ref: '#/components/schemas/SearchResult'
            suggestions:
              type: array
              items:
                type: string
            total:
              type: integer
            query:
              type: string
            mode:
              type: string
    SearchResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          format: uri
        relevance_score:
          type: number
          minimum: 0
          maximum: 1
        highlighted_name:
          type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Invalid request parameters"
            details:
              - field: "email"
                message: "Invalid email format"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Internal server error"
...